fix: test failures
