+++ packages/react-hooks/jest.config.ts
@@ -1,11 +1,23 @@
 import type { Config } from '@jest/types'
 
-import base from '../../jest.config.base'
-
 import packageJson from './package.json'
 
 const config: Config.InitialOptions = {
-  ...base,
+  testTimeout: 120000,
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  coveragePathIgnorePatterns: ['/build/', '/node_modules/', '/__tests__/', 'tests'],
+  coverageDirectory: '<rootDir>/coverage/',
+  verbose: true,
+  testMatch: ['**/?(*.)+(spec|test).[tj]s?(x)'],
+  transform: {
+    '^.+\\.tsx?$': [
+      'ts-jest',
+      {
+        isolatedModules: true,
+      },
+    ],
+  },
   displayName: packageJson.name,
 }
